% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_fourier_series}
\alias{compute_fourier_series}
\title{Compute Fourier Series}
\usage{
compute_fourier_series(mat, t, type)
}
\arguments{
\item{mat}{Data matrix of size nxp}

\item{t}{Vector of size m where the series is being evaluated.}

\item{type}{Type of curve.}
}
\value{
A nxm matrix where each row represents a curve.
}
\description{
Computes Fourier series required by Andrews curves.
}
