% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drewcurves.R
\name{drewcurves}
\alias{drewcurves}
\title{Andrews Plot}
\usage{
drewcurves(df, type = 1, group = NULL, resolution = 100,
  return_dataframe = FALSE, minmax_scaling = TRUE, xlim = c(-pi, pi))
}
\arguments{
\item{df}{An nxp dataframe or matrix.}

\item{type}{Type of Fourier series to be computed. See vignette.}

\item{group}{Integer indicating the number of column that
represents the group, if any.}

\item{resolution}{The number of regularly spaced points where the
Fourier series will be evaluated.}

\item{return_dataframe}{If TRUE, the function will return the
dataframe used to generate the Andrews plot rather than the
plot itself.}

\item{minmax_scaling}{If TRUE, it substract the mimimum and divide
by the length of the range of every column.}

\item{xlim}{A vector of size two with the limits where the Fourier
series will be evaluated. The default limits are are -pi and
pi.}
}
\value{
A ggplot object or a dataframe (depending on the parameter
    'return_dataframe').
}
\description{
This function computes the Andrews plot of a dataframe.
}
\details{
The default output is the Andrews plot generated with ggplot2.
}
\author{
Guillermo Basulto-Elias
}
