% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minmax}
\alias{minmax}
\title{Minmax Normalization}
\usage{
minmax(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
The normalized matrix, with a vector with ' column
minimums and a vector with the column range lengths as attributes.
}
\description{
For each column of a matrix, it substracts its minimum and divide
by the length of its range
}
\details{
An error is returned if there is at least one column whose values
are all the same.
}
\examples{
mat <- matrix(c(0, 1, 2,
                3, 4, 5,
                3, 2, 1),
                nrow = 3, byrow = TRUE)
  drewcurves:::minmax(mat)
  
\dontrun{
mat <- matrix(c(0, pi, 1, 2,
                3, pi, 4, 5,
                3, pi, 2, 1),
                nrow = 3, byrow = TRUE)
drewcurves:::minmax(mat)
  }
}
\author{
Guillermo Basulto-Elias
}
