% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax_normalization.R
\name{minmax_normalization}
\alias{minmax_normalization}
\title{Min-max Normalization}
\usage{
minmax_normalization(X)
}
\arguments{
\item{X}{Vector, matriz or dataframe.}
}
\value{
An object of the same type and dimensions as X.
}
\description{
Normalize columns by substracting its minimum and dividing by the
length of its range. The resulting columns will be between zero
and one. It does not affect non-numeric columns.
}
\examples{
data(iris)
minmax_normalization(iris[, -5])
}
\author{
Guillermo Basulto-Elias
}
